+++ src/access/isAdmin.ts
@@ -6,6 +6,6 @@ export const isAdmin: Access<any, User> = ({ req: { user } }) => {
   return Boolean(user?.roles?.includes('admin'));
 };
 
-export const isAdminFieldLevel: FieldAccess<{ id: string }, unknown, User> = ({ req: { user } }) => {
+export const isAdminField: FieldAccess<{ id: string }, unknown, User> = ({ req: { user } }) => {
   return Boolean(user?.roles?.includes('admin'));
 };