+++ src/access/isProtected.ts
@@ -1,0 +1,6 @@
+import { FieldAccess } from 'payload/types';
+
+import { User } from '../payload-types';
+
+export const isProtectedField: FieldAccess<any, User> = async ({ req: { user }, doc }) =>
+  user ? true : !!doc?.protected === false;